# web/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Rempart DNS",
    "version": "1.1",
    "description": "API REST de Rempart. Session (cookie rempart_session + en-tête X-Rempart: 1 sur les requêtes qui modifient) ou jeton porteur rmp_…. Les champs inconnus sont refusés (400)."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/login/otp": {
      "post": {
        "tags": [
          "Authentification"
        ],
        "summary": "Compléter la connexion par un code TOTP ou de secours",
        "description": "Portée : public — Sans authentification.",
        "operationId": "post_api_login_otp",
        "x-rempart-scope": "public",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "recovery_left": {
                      "type": "integer"
                    }
                  }
                },
                "example": {
                  "user": "admin",
                  "method": "totp",
                  "recovery_left": 10
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "challenge": {
                    "type": "string"
                  },
                  "code": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "challenge": "Q3h…",
                "code": "123456"
              }
            }
          }
        }
      }
    },
...
    "/api/otp": {
      "get": {
        "tags": [
          "Compte local : mot de passe, TOTP, clés d'accès"
        ],
        "summary": "État de la double authentification",
        "description": "Portée : local — Session du compte administrateur local seulement (ni jeton, ni compte LDAP/OIDC).",
        "operationId": "get_api_otp",
        "x-rempart-scope": "local",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Succès",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "enabled": {
                      "type": "boolean"
                    },
                    "recovery_left": {
                      "type": "integer"
                    }
                  }
                },
                "example": {
                  "enabled": false,
                  "recovery_left": 0
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "description": "Jeton rmp_… créé dans Sécurité → Accès API ou par POST /api/tokens."
      },
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "rempart_session",
        "description": "Cookie posé par POST /api/login."
      }
    },
    "responses": {
      "Error": {
        "description": "Erreur",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  }
}